Add initial code for connecting to web service
This makes use of two dependencies for receiving/transmitting Parameters from the web service. This is all very preliminary code so use it more as a suggestion than a concrete implementation.

ParametersWrapper is a singleton that directs the model to either use our webservice or the bundled XML. At the moment this is done with a simple boolean switch, but this could easily be made aware of the endpoint and default to XML if the webservice timed out for example.

WebclientParameters is an implementation of Repast’s Parameters interface which will communicate with the backend. Most of the methods are just stubs that will need to be fully fleshed out. Eventually we will want to pass all the information from the XML from the web service, including the metadata like “displayName”, “isReadOnly”, and “type”.

To make our life easier connecting to the REST backend we use Unirest, a lightweight client that wraps a bunch of Apache HTTP classes under the hood. The added Jar has all its dependencies added in to one fat Jar.

To make dealing with JSON easier we use Jackson Jr, a simplified and lightweight implementation of Jackson which should be more than enough for this simple web service. The added Jar has all its dependencies added into one fat jar.